clang implementation

+++ Interpreter_Cyberdes/lib/Interpreter.cpp
@@ -23,7 +23,7 @@ void Interpreter::runCode(const std::vector<std::string>& code) {
             case PUSH:
                 programCounter++;
                 if (isNumber(code[programCounter])) {
-                    stack.push_back(std::stof(code[programCounter]));
+                    stack.push_back(std::stoi(code[programCounter]));
                 } else {
                     stack.push_back(code[programCounter]);
                 }
